+++ x/distribution/migrations/v1/types.go
@@ -5,7 +5,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/kv"
-	v1auth "github.com/cosmos/cosmos-sdk/x/auth/migrations/v1"
 )
 
 const (
@@ -20,6 +19,9 @@ const (
 
 	// QuerierRoute is the querier route for distribution
 	QuerierRoute = ModuleName
+
+	addrLen        = 20
+	authModuleName = "auth"
 )
 
 // Keys for distribution store
@@ -59,7 +61,7 @@ var (
 func GetValidatorOutstandingRewardsAddress(key []byte) (valAddr sdk.ValAddress) {
 	kv.AssertKeyAtLeastLength(key, 2)
 	addr := key[1:]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyLength(addr, addrLen)
 	return sdk.ValAddress(addr)
 }
 
@@ -67,29 +69,29 @@ func GetValidatorOutstandingRewardsAddress(key []byte) (valAddr sdk.ValAddress) 
 func GetDelegatorWithdrawInfoAddress(key []byte) (delAddr sdk.AccAddress) {
 	kv.AssertKeyAtLeastLength(key, 2)
 	addr := key[1:]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyLength(addr, addrLen)
 	return sdk.AccAddress(addr)
 }
 
 // gets the addresses from a delegator starting info key
 func GetDelegatorStartingInfoAddresses(key []byte) (valAddr sdk.ValAddress, delAddr sdk.AccAddress) {
-	kv.AssertKeyAtLeastLength(key, 2+v1auth.AddrLen)
-	addr := key[1 : 1+v1auth.AddrLen]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyAtLeastLength(key, 2+addrLen)
+	addr := key[1 : 1+addrLen]
+	kv.AssertKeyLength(addr, addrLen)
 	valAddr = sdk.ValAddress(addr)
-	addr = key[1+v1auth.AddrLen:]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	addr = key[1+addrLen:]
+	kv.AssertKeyLength(addr, addrLen)
 	delAddr = sdk.AccAddress(addr)
 	return
 }
 
 // gets the address & period from a validator's historical rewards key
 func GetValidatorHistoricalRewardsAddressPeriod(key []byte) (valAddr sdk.ValAddress, period uint64) {
-	kv.AssertKeyAtLeastLength(key, 2+v1auth.AddrLen)
-	addr := key[1 : 1+v1auth.AddrLen]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyAtLeastLength(key, 2+addrLen)
+	addr := key[1 : 1+addrLen]
+	kv.AssertKeyLength(addr, addrLen)
 	valAddr = sdk.ValAddress(addr)
-	b := key[1+v1auth.AddrLen:]
+	b := key[1+addrLen:]
 	kv.AssertKeyLength(addr, 8)
 	period = binary.LittleEndian.Uint64(b)
 	return
@@ -99,7 +101,7 @@ func GetValidatorHistoricalRewardsAddressPeriod(key []byte) (valAddr sdk.ValAddr
 func GetValidatorCurrentRewardsAddress(key []byte) (valAddr sdk.ValAddress) {
 	kv.AssertKeyAtLeastLength(key, 2)
 	addr := key[1:]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyLength(addr, addrLen)
 	return sdk.ValAddress(addr)
 }
 
@@ -107,17 +109,17 @@ func GetValidatorCurrentRewardsAddress(key []byte) (valAddr sdk.ValAddress) {
 func GetValidatorAccumulatedCommissionAddress(key []byte) (valAddr sdk.ValAddress) {
 	kv.AssertKeyAtLeastLength(key, 2)
 	addr := key[1:]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyLength(addr, addrLen)
 	return sdk.ValAddress(addr)
 }
 
 // gets the height from a validator's slash event key
 func GetValidatorSlashEventAddressHeight(key []byte) (valAddr sdk.ValAddress, height uint64) {
-	kv.AssertKeyAtLeastLength(key, 2+v1auth.AddrLen)
-	addr := key[1 : 1+v1auth.AddrLen]
-	kv.AssertKeyLength(addr, v1auth.AddrLen)
+	kv.AssertKeyAtLeastLength(key, 2+addrLen)
+	addr := key[1 : 1+addrLen]
+	kv.AssertKeyLength(addr, addrLen)
 	valAddr = sdk.ValAddress(addr)
-	startB := 1 + v1auth.AddrLen
+	startB := 1 + addrLen
 	kv.AssertKeyAtLeastLength(key, startB+9)
 	b := key[startB : startB+8] // the next 8 bytes represent the height
 	height = binary.BigEndian.Uint64(b)